fix sync-blog script
