--- tsconfig.cli.json
@@ -1,6 +1,0 @@
-{
-  "extends": "./tsconfig",
-  "compilerOptions": {
-    "module": "commonjs"
-  }
-}
